:white_check_mark::technologist: add mypy lint in CI

+++ docker/v3.10/entrypoints/test.sh
@@ -1,18 +1,43 @@
 # /bin/bash
 
+testReturn=0
+
+failed() {
+    testReturn=1
+    echo "FAILED $1"
+}
+
+set -x
+
 # run the tests
 # -- lint
-poetry run flake8 --output-file flake8.txt || echo "FAILED flake"
-poetry run flake8_junit flake8.txt flake8.xml
-rm flake8.txt
+if ! poetry run flake8 --output-file flake8.txt
+then
+    failed "flake";
+    poetry run flake8_junit flake8.txt flake8.xml && echo;
+    rm flake8.txt;
+fi
+
+# -- type checking
+if ! poetry run mypy --junit-xml ./mypy.xml .
+then
+    failed "mypy";
+fi
 
 # -- unit test & coverage
 # -- api
-poetry run pytest --junitxml=pytest.xml --cov --html=pytest.html
-poetry run coverage xml
+if poetry run pytest --junitxml=pytest.xml --cov --html=pytest.html
+then
+    poetry run coverage xml
+else
+    failed "pytest"
+fi
 
 # -- build package
-poetry build
+if ! poetry build
+then
+    failed "build"
+fi
 
 # prepare test results
 echo "Moving test results file..."
@@ -20,3 +45,5 @@ mkdir -p testResults
 mv *.xml  ./testResults
 mv *.html  ./testResults
 echo "Done"
+
+exit $testReturn
